+++ CHANGELOG.md
@@ -55,6 +55,7 @@ This name should be decided amongst the team before the release.
 - [#1254](https://github.com/topiary/topiary/pull/1254) Topiary now gracefully skips formatting for unknown or unsupported injected languages instead of returning a fatal error, allowing files with unconfigured code blocks (like Markdown) to format successfully.
 - [#1267](https://github.com/topiary/topiary/pull/1267) Add initial injection support (JSON) for Rust
 - [#1270](https://github.com/topiary/topiary/pull/1270) Add initial `languages.ncl` nickel contracts
+- [#1139](https://github.com/topiary/topiary/pull/1139) Added `rootcause::Report` handling in topiary-cli.
 
 ### Fixed
 - [#1176](https://github.com/topiary/topiary/pull/1176) Increase the stack size to 4MiB in Windows builds.